Tests/reorganize: move error recording test to own file
